Update headers after restructure
